refactor: split plutus encoder into focused package files
